+++ src/utils/mapbox/index.ts
@@ -2,4 +2,4 @@
 export * from './constants';
 export * from './types';
 export * from './api';
-export * from './helpers';
+export * from './misc';
